latest changes
not worked on this for a while so I forget what changed

+++ src/views/partials/role_group_show.blade.php
@@ -38,20 +38,41 @@
                         @endif
                     </p>
                     <hr/>
-                    @if (Auth::user()->hasRole('role_group_edit'))
+                    @if (!$role_group->trashed() && Auth::user()->hasRole('role_group_edit'))
                         <a href="{{ route('role_groups.edit', $role_group->id) }}" class="btn btn-primary">
                             <span class="glyphicon glyphicon-edit" aria-hidden="true"></span>
                             <span>Edit</span>
                         </a>
                     @endif
-                    <a class="btn btn-default" href="{{ route('role_groups.index') }}">Browse Role Groups</a>
-                    @if (Auth::user()->hasRole('role_group_delete'))
+                    @if (Auth::user()->hasRole('role_group_browse'))
+                        <a class="btn btn-default" href="{{ route('role_groups.index') }}">
+                            <span>Browse Role Groups</span>
+                        </a>
+                    @endif
+                    @if (!$role_group->trashed() && Auth::user()->hasRole('role_group_delete'))
                         {!! Form::open(['url' => route('role_groups.destroy', $role_group->id), 'method' => 'DELETE', 'class' => 'form-horizontal pull-right', 'role' => 'form']) !!}
-                        <button type="submit" class="btn btn-danger">
-                            <span class="glyphicon glyphicon-trash" aria-hidden="true"></span>
-                            <span>Delete</span>
-                        </button>
+                            <button type="submit" class="btn btn-danger">
+                                <span class="glyphicon glyphicon-trash" aria-hidden="true"></span>
+                                <span>Delete</span>
+                            </button>
                         {!! Form::close() !!}
+                    @elseif ($role_group->trashed())
+                        @if (Auth::user()->hasRole('role_group_destroy'))
+                            {!! Form::open(['url' => route('role_groups.destroy', $role_group->id), 'method' => 'DELETE', 'class' => 'form-horizontal pull-right', 'role' => 'form']) !!}
+                            <button type="submit" class="btn btn-danger">
+                                <span class="glyphicon glyphicon-trash" aria-hidden="true"></span>
+                                <span>Destroy</span>
+                            </button>
+                        {!! Form::close() !!}
+                        @endif
+                        @if (Auth::user()->hasRole('role_group_restore'))
+                            {!! Form::open(['url' => route('role_groups.update', $role_group->id), 'method' => 'PUT', 'class' => 'form-horizontal pull-right', 'role' => 'form', 'style' => 'margin-right:4px;']) !!}
+                                <button type="submit" class="btn btn-primary">
+                                    <span class="glyphicon glyphicon-repeat" aria-hidden="true"></span>
+                                    <span>Restore</span>
+                                </button>
+                            {!! Form::close() !!}
+                        @endif
                     @endif
                 </div>
             </div>
